backend/letter_soup/pyproject.toml
[project]
name = "letter-soup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy",
    "pydantic",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "python-multipart",
    "pydantic-settings",
    "psycopg[binary]",
    "psycopg2-binary",
    "sqlalchemy-utils",
  



]

[project.scripts]
start = "main:main"

[tool.setuptools]
packages = ["src"]
py-modules = ["main"]
include-package-data = true


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"